Added Mock in Wompi Payment.
